producer/requirements.txt
kafka-python
twitchio
requests